requirements.txt
# Основные зависимости для SQLite-версии приложения
streamlit>=1.22.0
pandas>=1.5.3
folium>=0.14.0
streamlit-folium>=0.11.0
plotly>=5.13.0
numpy>=1.24.3

# Зависимости для экспорта данных
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0

# Зависимости для планировщика
schedule>=1.2.0
gitpython>=3.1.30

# Дополнительные зависимости
requests>=2.28.1 